move sysctl code into separate module
This should make the code easier to work with. I will add more
functionallity later.
